move test utils to own file
