+++ test/index.js
@@ -1,35 +1,7 @@
 const fs = require('fs');
 const test = require('tape');
 const globrex = require('../');
-const isWin = process.platform === 'win32';
-const GLOBSTAR = isWin ? '((?:[^\\\\]*(?:\\\\|$))*)' : '((?:[^\\/]*(?:\\/|$))*)';
-
-// TODO: Rewrite tests match(t, { unix, win, expect, opts})
-function match(glob, strUnix, strWin, opts = {}) {
-   if (typeof strWin === 'object') {
-      opts = strWin;
-      strWin = false;
-   }
-   let res = globrex(glob, opts);
-   return res.regex.test(isWin && strWin ? strWin : strUnix);
-}
-
-// TODO: To string - take real regex to avoid escape hell
-function matchRegex(t, pattern, ifUnix, ifWin, opts) {
-   const res = globrex(pattern, opts);
-   const {regex} = (opts.filepath ? res.path : res);
-   t.is(regex.toString(), isWin ? ifWin : ifUnix, '~> regex matches expectant');
-   return res;
-}
-
-// TODO: Move to utils
-function matchSegments(t, pattern, ifUnix, ifWin, opts) {
-   const res = globrex(pattern, {filepath:true, ...opts});
-   const str = res.path.segments.join(' ');
-   const exp = (isWin ? ifWin : ifUnix).join(' ');
-   t.is(str, exp);
-   return res;
-}
+const {match, matchRegex, matchSegments} = require('./helpers');
 
 test('globrex: standard', t => {
    t.plan(3);
